desktopapplications: watch files/folders for updates
* directory and file watching for desktop entries

* Removed paranoid mutex'ing

* expanded the concurrency paranoia

* remembered to make filesystem change handling async

* feat: added 'send' command to send socket requests

* consolidated watcher function and added concurrency protection

Moved the guts of the startWatcher() function into loadFiles().\nDoing so reduces the amount of disk operations and should reduce the amount of time loading.\nShould only be noticable on especially large watchsets\n\nAdded a mutex to the watchedDirs mapping, colissions are unlikely given the nature of them but during setup walkers could collide, leading to a startup panic.

* Continued mutex paranoia, added support for symlinked desktop entries

There was a few changes and additions in this one. We now use a two-way mapping system to deal with symlinks and their nuance. Filesystem watching was reworked a little to account for some of the design implications that comes from support symlinks as well. loadFiles() saw some abstraction, namely the introduction of the setVars() function that handles the oneshot var setup, and the abstraction of walkFunction() which was moved out of the load scope to improve readability. Some functions, such as the specific event handlers, now have rudimentary logic documentation (sorry).

---------

Co-authored-by: Andrej Benz <hello@benz.dev>